US-08 - added validators for backend
